+++ docs/changes.rst
@@ -330,6 +330,11 @@ Released: 2019-07-20
 * Docs: Clarified how the pywbem_os_setup.sh/bat scripts can be downloaded
   using a predictable URL, for automated downloads.
 
+* Clarified the 'x509' parameter of 'WBEMConnection' in that its 'key_file'
+  item is optional and if omitted, both the private key and the certificate
+  must be in the file referenced by the 'cert_file' item. Added checks
+  for the 'x509' parameter.
+
 **Cleanup:**
 
 * Test: Removed pinning of distro version on Travis to Ubuntu xenial (16.04)